+++ CMakeLists.txt
@@ -109,7 +109,7 @@ set(T_COSE_SRC_COMMON
     src/t_cose_recipient_dec_esdh.c
     src/t_cose_recipient_enc_esdh.c
     src/t_cose_qcbor_gap.c
-    src/t_cose_private_test.c
+    src/t_cose_private.c
 )
 
 # Some functions in QCBOR require the math library, -lm, but t_cose
