+++ Makefile.common
@@ -35,7 +35,7 @@ SRC_OBJ=src/t_cose_util.o \
         src/t_cose_recipient_enc_keywrap.o \
         src/t_cose_recipient_dec_esdh.o \
         src/t_cose_recipient_enc_esdh.o \
-        src/t_cose_private_test.o \
+        src/t_cose_private.o \
         src/t_cose_qcbor_gap.o
